requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
gcloud==0.17.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
jws==0.1.3
oauth2client==3.0.0
paho-mqtt==1.5.1
protobuf==4.0.0rc2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
whitenoise==5.2.0
